fixup! feature(custom-option): Add query in custom option
