+++ app/views/EditProject/UpdateProjectForm/ValidateProjectSpecifics/index.tsx
@@ -10,6 +10,7 @@ import {
     useFormObject,
 } from '@togglecorp/toggle-form';
 import { ulid } from 'ulid';
+import { gql } from 'urql';
 
 import Button from '#components/Button';
 import Container from '#components/Container';
@@ -21,7 +22,10 @@ import {
     type PartialRasterTileServerInputFields,
 } from '#components/domain/RasterTileServerInput/schema';
 import NonFieldError from '#components/NonFieldError';
-import { ProjectTypeEnum } from '#generated/types/graphql';
+import {
+    ProjectTypeEnum,
+    useDefaultCustomOptionsQuery,
+} from '#generated/types/graphql';
 
 import {
     defaultObjectSourceInputFormValue,
@@ -29,6 +33,19 @@ import {
 } from './ObjectSourceInput/schema';
 import ObjectSourceInput from './ObjectSourceInput';
 import { type PartialValidateSpecificFields } from './schema';
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+const DEFAULT_CUSTOM_OPTIONS = gql`
+query DefaultCustomOptions($projectType: ProjectTypeEnum!) {
+    defaultCustomOptions(projectType: $projectType) {
+        value
+        title
+        iconColor
+        icon
+        description
+    }
+}
+`;
 
 interface Props {
     projectId: string;
@@ -48,6 +65,14 @@ function ValidateProjectSpecifics(props: Props) {
         disabled,
         projectType,
     } = props;
+
+    const [
+        { data: customOptionResponse },
+    ] = useDefaultCustomOptionsQuery({
+        variables: {
+            projectType,
+        },
+    });
 
     const error = getErrorObject(formError);
 
@@ -71,10 +96,16 @@ function ValidateProjectSpecifics(props: Props) {
         setFieldValue,
     );
 
-    const addCustomOption = useCallback((newCustomOptionIndex: number) => {
+    const addCustomOption = useCallback((index: number) => {
+        const defaultOption = customOptionResponse?.defaultCustomOptions?.[index];
+
         const newCustomOption: PartialCustomOptionInputFields = {
             clientId: ulid(),
-            value: newCustomOptionIndex,
+            value: defaultOption?.value ?? index,
+            title: defaultOption?.title,
+            icon: defaultOption?.icon,
+            iconColor: defaultOption?.iconColor,
+            description: defaultOption?.description,
         };
 
         setFieldValue(
@@ -83,7 +114,7 @@ function ValidateProjectSpecifics(props: Props) {
             ),
             'customOptions' as const,
         );
-    }, [setFieldValue]);
+    }, [setFieldValue, customOptionResponse]);
 
     return (
         <>
@@ -115,7 +146,6 @@ function ValidateProjectSpecifics(props: Props) {
                         index={optionIndex}
                         value={customOption}
                         onChange={setCustomOptionValue}
-                        projectType={projectType}
                         error={getErrorObject(
                             getErrorObject(error?.customOptions)?.[customOption.clientId],
                         )}
